Move MainContent components
