+++ packages/lib-user/src/components/shared/MainContent/components/StyledButton.js
@@ -1,0 +1,10 @@
+import { Button } from 'grommet'
+import styled from 'styled-components'
+
+const StyledButton = styled(Button)`
+  background-color: ${props => props.theme.global.colors['neutral-1']};
+  border-radius: 4px;
+  color: ${props => props.theme.global.colors['neutral-6']};
+`
+
+export default StyledButton
