+++ packages/lib-user/src/components/shared/MainContent/components/index.js
@@ -1,0 +1,2 @@
+export { default as StyledButton } from './StyledButton'
+export { default as StyledTab } from './StyledTab'
